refactor(mocks): extract boards module from mockApi (#991 M3)
